% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreedingSchemeLanguage.R
\docType{package}
\name{BreedingSchemeLanguage}
\alias{BreedingSchemeLanguage}
\alias{BreedingSchemeLanguage-package}
\title{BreedingSchemeLanguage}
\description{
Describe and simulate breeding schemes
}

