% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationFunc.R
\name{calcGenotypicValue}
\alias{calcGenotypicValue}
\title{calcGenotypicValue}
\usage{
calcGenotypicValue(geno, mapData)
}
\arguments{
\item{geno}{matrix of haplotypes}

\item{mapData}{map data}
}
\description{
calcGenotypicValue
}

