% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationFunc.R
\name{calcPhenotypicValue}
\alias{calcPhenotypicValue}
\title{calcPhenotypicValue}
\usage{
calcPhenotypicValue(gv, errorVar = NULL, H2 = NULL)
}
\arguments{
\item{gv}{genotypic values}

\item{errorVar}{error variance}

\item{H2}{a broad heritability}
}
\description{
calcPhenotypicValue
}

