% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{Random mate}
\usage{
cross(nProgeny = 100, equalContribution = F, popID = NULL)
}
\arguments{
\item{nProgeny}{the number of progenies}

\item{equalContribution}{if T all individuals used the same number of times as parents, if F individuals chosen at random to be parents}

\item{popID}{population ID to be crossed (default: the latest population)}
}
\value{
sequence information of progenies and the all information created before (list)
}
\description{
Random mate
}

