% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledHaploid.R
\name{doubledHaploid}
\alias{doubledHaploid}
\title{Doubled haploids}
\usage{
doubledHaploid(nProgeny = 100, popID = NULL)
}
\arguments{
\item{nProgeny}{the number of progenies}

\item{popID}{population ID to be devided by meiosis and doubled (default: the latest population)}
}
\value{
sequence information of progenies and the all information created before (list)
}
\description{
Doubled haploids
}

