% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{genotype}
\alias{genotype}
\title{Genotype markers}
\usage{
genotype(simEnv, popID = NULL)
}
\arguments{
\item{simEnv}{an environment that BSL statements operate on}

\item{popID}{population ID to be genotyped (default: the latest population)}
}
\value{
marker genotype and the all information created before (list)
}
\description{
Genotype markers
}

