% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoalescentSim.R
\name{getCoalescentSim}
\alias{getCoalescentSim}
\title{getCalescentSim}
\usage{
getCoalescentSim(nPopsSamples = NULL, effPopSize = 100, nChr = 7,
  nPiecesPerChr = 15000, recBTpieces = 1e-04, nMrkOrMut = 100,
  minMAF = 0.01, seed = as.integer(Sys.time()), tree = 0)
}
\arguments{
\item{nPopsSamples}{-pop[2] (GENOME)}

\item{effPopSize}{-N (GENOME)}

\item{nChr}{-c (GENOME)}

\item{nPiecesPerChr}{-pieces (GENOME)}

\item{recBTpieces}{-rec (GENOME)}

\item{nMrkOrMut}{parameter tocalculate -s (GENOME)}

\item{minMAF}{-maf (GENOME)}

\item{seed}{-seed (GENOME)}

\item{tree}{-tree (GENOME)}
}
\description{
getCalescentSim
}

