% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Import phased data from hapmap format file defining founder haplotypes}
\usage{
importData(fileName = "founderHaplotypes.hmp")
}
\arguments{
\item{fileName}{if null read a file called "founderHaplotypes.hmp" (default) else read
the specified file.}
}
\description{
Import phased data from hapmap format file defining founder haplotypes
}

