% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinBlup.R
\name{kin.blup4.4}
\alias{kin.blup4.4}
\title{kin.blup4.4}
\usage{
kin.blup4.4(data, geno, pheno, GAUSS = FALSE, K = NULL, fixed = NULL,
  covariate = NULL, PEV = FALSE, n.core = 1, theta.seq = NULL,
  reduce = FALSE, R = NULL)
}
\arguments{
\item{data}{see "rrBLUP"}

\item{geno}{see "rrBLUP"}

\item{pheno}{see "rrBLUP"}

\item{GAUSS}{see "rrBLUP"}

\item{K}{see "rrBLUP"}

\item{fixed}{see "rrBLUP"}

\item{covariate}{see "rrBLUP"}

\item{PEV}{see "rrBLUP"}

\item{n.core}{see "rrBLUP"}

\item{theta.seq}{see "rrBLUP"}

\item{reduce}{see "rrBLUP"}

\item{R}{see "rrBLUP"}
}
\description{
kin.blup4.4
}

