% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMap.R
\name{makeMap}
\alias{makeMap}
\title{create map and QTL effects}
\usage{
makeMap(map, nLoci, nMarkers, nQTL, propDomi, interactionMean, varEffects = 1)
}
\arguments{
\item{map}{map information (Chromosome and Position)}

\item{nLoci}{the number of markers and QTL}

\item{nMarkers}{the number of markers, which is used especially for genomic selection}

\item{nQTL}{the number of QTLs controlling the target trait}

\item{propDomi}{the probability of dominant QTL among the all QTL}

\item{interactionMean}{the expected number of epistatic loci for each effect}

\item{varEffects}{variance of QTL effects}
}
\value{
The simulation results (The output data was saved as BSLoutput.RData. After you load the data in R, you can find the data named as BSLoutput.)
}
\description{
create map and QTL effects
}

