% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossingFunc.R
\name{makeProgenies}
\alias{makeProgenies}
\title{makeProgenies}
\usage{
makeProgenies(parents, geno, pos)
}
\arguments{
\item{parents}{ID of haplotypes that the parents harbor}

\item{geno}{matrix of haplotypes}

\item{pos}{position of markers/QTLs}
}
\description{
makeProgenies
}

