% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossingFunc.R
\name{makeProgeny}
\alias{makeProgeny}
\title{makeProgeny}
\usage{
makeProgeny(genoMat, genoPat, pos)
}
\arguments{
\item{genoMat}{matrix of maternal haplotype}

\item{genoPat}{matrix of paternal haplotype}

\item{pos}{position of markers/QTLs}
}
\description{
makeProgeny
}

