% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossingFunc.R
\name{pedigreeMate}
\alias{pedigreeMate}
\title{pedigreeMate}
\usage{
pedigreeMate(parents, geno, pos)
}
\arguments{
\item{parents}{two- or three-column matrix: first two columns index the parents you want to cross, if there is a third column, it is the number of progeny from that pair of parents}

\item{geno}{matrix of haplotypes}

\item{pos}{position of markers/QTLs}
}
\description{
pedigreeMate
}
