% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{phasedHapMap2mat}
\alias{phasedHapMap2mat}
\title{Transform a data.frame with a hapmap data in it into a marker dosage and map list}
\usage{
phasedHapMap2mat(hm, firstCol = 12, lastCol = ncol(hm))
}
\arguments{
\item{hm}{The data.frame}

\item{firstCol}{Number of the first column you want in the marker dosage file}

\item{lastCol}{Number of the last column you want in the marker dosage file}
}
\value{
list with markers and map objects
}
\description{
Transform a data.frame with a hapmap data in it into a marker dosage and map list
}

