% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Evaluate the phenotypic value}
\usage{
phenotype(errorVar = 1, popID = NULL)
}
\arguments{
\item{errorVar}{error variance}

\item{popID}{population ID to be evaluated (default: the latest population)}
}
\value{
phenotypic values and the all information created before (list)
}
\description{
Evaluate the phenotypic value
}

