% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossingFunc.R
\name{predGameteMeanVar}
\alias{predGameteMeanVar}
\title{predGameteMeanVar}
\usage{
predGameteMeanVar(geno, pos, locEff)
}
\arguments{
\item{geno}{2 x nLoc matrix of haplotypes, coded -1 and 1}

\item{pos}{position of markers/QTLs}

\item{locEff}{effects of each locus}
}
\description{
predGameteMeanVar
}
