% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictBreedVal}
\alias{predictBreedVal}
\title{Genomic prediction}
\usage{
predictBreedVal(popID = NULL, trainingPopID = NULL)
}
\arguments{
\item{popID}{population ID to be predicted (default: the latest population)}

\item{trainingPopID}{population ID to be used for training a prediction model (default: all populations with phenotype data)}
}
\value{
predicted values and the all information created before (list)
}
\description{
Genomic prediction
}

