% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossingFunc.R
\name{randomMate}
\alias{randomMate}
\title{randomMate}
\usage{
randomMate(popSize, geno, pos)
}
\arguments{
\item{popSize}{population size}

\item{geno}{matrix of haplotypes}

\item{pos}{position of markers/QTLs}
}
\description{
randomMate
}

