% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossingFunc.R
\name{randomMateNoFam}
\alias{randomMateNoFam}
\title{randomMateNoFam}
\usage{
randomMateNoFam(popSize, geno, pos, genoRec)
}
\arguments{
\item{popSize}{the number of progeny to return}

\item{geno}{matrix of haplotypes}

\item{pos}{position of markers/QTLs}
}
\description{
randomMateNoFam
}
