% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select individuals}
\usage{
select(nSelect = 40, popID = NULL, random = F)
}
\arguments{
\item{nSelect}{the number of selected individuals}

\item{popID}{population ID to be selected (default: When random=T, the last population. When random=F, it is the last evaluated population)}

\item{random}{assuming random selection or selection according to their features (T: random selection, F: selection of good individuals)}
}
\value{
information of the selected individuals and the all information created before (list)
}
\description{
Select individuals
}

