% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfFertilize.R
\name{selfFertilize}
\alias{selfFertilize}
\title{Self-fertilize}
\usage{
selfFertilize(simEnv, nProgenyPerInd = 1, popID = NULL)
}
\arguments{
\item{popID}{population ID to be self-fertilized (default: the latest population)}

\item{nProgeny}{the number of progeny}
}
\value{
sequence information of progenies and the all information created before (list)
}
\description{
Self-fertilize
}

