% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testParameterOptimality.R
\name{testParameterOptimality}
\alias{testParameterOptimality}
\title{Function to return the optimality of a parameter vector for a breeding scheme given a simulation environment}
\usage{
testParameterOptimality(sEnv = NULL, schemeFileName, parmVec, budget)
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. This environment should be fresh from defineSpecies, defineVariances, and defineCosts. If NULL, the default \code{simEnv} is attempted}

\item{schemeFileName}{source file that holds the script of the breeding scheme. The scheme should return a value to be maximized in a variable called \code{objective}}

\item{parmVec}{vector with values of parameters characterizing the breeding scheme}

\item{budget}{the maximum budget that is allowed for the breeding scheme}
}
\value{
Optimality of the breeding scheme given the simulation environment and the parameter vector. If parameter vector causes the scheme to exceed the budget, NA is returned
}
\description{
Function to return the optimality of a parameter vector for a breeding scheme given a simulation environment
}
