% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{Cross with random mating, or equal contributions, or randomly between two populations}
\usage{
cross(sEnv = NULL, nProgeny = 100, equalContribution = F, popID = NULL,
  popID2 = NULL, notWithinFam = F)
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{nProgeny}{the number of progenies}

\item{equalContribution}{if T all individuals used the same number of times as parents, if F individuals chosen at random to be parents}

\item{popID}{population ID to be crossed. Default: the last population}

\item{popID2}{population ID to be crossed with popID to make hybrids. Default: second population not used}

\item{notWithinFam}{if TRUE, like equalContribution, all individuals are used the same number of times as parents and self-fertilization is not allowed. In addition, half- and full-sibs are not allowed to mate}
}
\value{
modifies the list sims in environment sEnv by creating a progeny population as specified, with an incremented population number
}
\description{
Cross with random mating, or equal contributions, or randomly between two populations
}
