% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubledHaploid.R
\name{doubledHaploid}
\alias{doubledHaploid}
\title{Doubled haploids}
\usage{
doubledHaploid(sEnv = simEnv, nProgeny = 100, popID = NULL)
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{nProgeny}{the number of progeny}

\item{popID}{population ID to be devided by meiosis and doubled (default: the latest population)}
}
\value{
modifies the list sims in environment sEnv by creating a doubled haploid progeny population as specified, with an incremented population number
}
\description{
Doubled haploids
}
