% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializePopulation.R
\name{initializePopulation}
\alias{initializePopulation}
\title{Create a founder population}
\usage{
initializePopulation(sEnv = NULL, nInd = 100)
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{nInd}{population size}
}
\value{
modifies the list sims in environment sEnv by creating a founder population
}
\description{
Create a founder population
}
