% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializePopulation.R
\name{initializePopulation}
\alias{initializePopulation}
\title{Create a founder population}
\usage{
initializePopulation(nPop = 100, gVariance = 1)
}
\arguments{
\item{nPop}{population size}

\item{gVariance}{genetic variance in the initial population}
}
\value{
initial population informationand the all information created before (list)
}
\description{
Create a founder population
}

