% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Evaluate the phenotypic value}
\usage{
phenotype(sEnv = simEnv, errorVar = 1, popID = NULL, locations = 1,
  years = 1)
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{errorVar}{error variance}

\item{popID}{population ID to be evaluated (default: the latest population)}

\item{locations}{integer vector of the locations where phenotyping occurs (e.g., c(1, 3) to phenotype at locations 1 and 3. Default: 1, phenotype at the first location)}

\item{years}{integer vector of the years when phenotypin occurs (e.g., 1:2 to phenotype during the first two years of the breeding scheme. Default: 1, phenotype the first year)}
}
\value{
phenotypic values and the all information created before (list)
}
\description{
Evaluate the phenotypic value
}

