% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossingFunc.R
\name{randomMateAll}
\alias{randomMateAll}
\title{randomMateAll}
\usage{
randomMateAll(popSize, geno, pos)
}
\arguments{
\item{popSize}{population size}

\item{geno}{matrix of haplotypes}

\item{pos}{position of markers/QTLs}
}
\description{
randomMateAll
}

