% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfFertilize.R
\name{selfFertilize}
\alias{selfFertilize}
\title{Self-fertilize}
\usage{
selfFertilize(simEnv, nProgeny = 100, popID = NULL)
}
\arguments{
\item{simEnv}{an environment that BSL statements operate on}

\item{nProgeny}{the number of progeny}

\item{popID}{population ID to be self-fertilized (default: the latest population)}
}
\value{
sequence information of progenies and the all information created before (list)
}
\description{
Self-fertilize
}

